requirements_test.txt
# Testing dependencies for Alfen Wallbox integration
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0

# Code quality tools
mypy>=1.8.0
ruff>=0.1.0
